pyproject.toml
[tool.poetry]
name = "betafrontrowcrew"
version = "0.1.0"
description = ""
authors = ["Scott Rubin <apreche@frontrowcrew.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
django-extensions = "^3.1.5"
celery = "^5.2.1"
gunicorn = "^20.1.0"
Pillow = "^9.0.0"
docutils = "^0.18.1"
psycopg2 = "^2.9.2"
pylibmc = "^1.6.1"
django-storages = "^1.12.3"
django-celery-results = "^2.2.0"
factory-boy = "^3.2.1"
Markdown = "^3.3.6"
mdgen = "^0.1.9"
boto3 = "^1.20.26"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
ipython = "^7.30.1"
ipdb = "^0.13.9"
Werkzeug = "^2.0.2"
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
